+++ instances/request_server.cpp
@@ -4,8 +4,8 @@
 #include <string.h>
 #include <unistd.h>
 
+#include <extras/rsi/subsystems.hpp>
 #include <extras/strings.hpp>
-#include <extras/uploader/UploaderInterface.hpp>
 #include <fstream>
 #include <iostream>
 #include <sstream>
@@ -35,7 +35,8 @@ int main(int argc, char const *argv[]) {
   //
   int sockfd;
   struct sockaddr_in server_addr;
-  sockfd = configure_serversocket(ip.c_str(), port, server_addr, false);
+  sockfd =
+      extras::rsi::configure_serversocket(ip.c_str(), port, server_addr, false);
 
   for (int i = 0; i < 1000; i++) {
     struct sockaddr_in new_addr;