+++ src/rsi/subsystem/send_file.cpp
@@ -4,18 +4,18 @@
 #include <string.h>
 #include <unistd.h>
 
-#include <extras/uploader/UploaderInterface.hpp>
+#include <extras/rsi/subsystems.hpp>
 
-void send_file(FILE *fp, int sockfd) {
+void extras::rsi::send_file(FILE *fp, int sockfd) {
   // int n;
-  char data[SIZE] = {0};
+  char data[extras::rsi::SIZE] = {0};
 
-  while (fgets(data, SIZE, fp) != NULL) {
+  while (fgets(data, extras::rsi::SIZE, fp) != NULL) {
     int len = strlen(data);
     if (send(sockfd, data, len, 0) == -1) {
       perror("[-]Error in sending file.");
       exit(1);
     }
-    bzero(data, SIZE);
+    bzero(data, extras::rsi::SIZE);
   }
 }