+++ src/sockets/Requests.cpp
@@ -1,6 +1,6 @@
 #include "extras/sockets/Requests.hpp"
 
-#include <extras/uploader/UploaderInterface.hpp>
+#include <extras/rsi/subsystems.hpp>
 #include <fstream>
 #include <iostream>
 #include <sstream>
@@ -12,8 +12,8 @@ namespace extras {
 
   PortNumber Requests::request(const RequestedService& serviceName,
                                const PortServerNumber& serverSocket) {
-    send_line(serviceName, serverSocket);
-    int port_to_use = stoi(read_line(serverSocket));
+    extras::rsi::send_line(serviceName, serverSocket);
+    int port_to_use = stoi(extras::rsi::read_line(serverSocket));
     return port_to_use;
   }