+++ src/sockets/Services.cpp
@@ -1,6 +1,6 @@
 #include "extras/sockets/Services.hpp"
 
-#include <extras/uploader/UploaderInterface.hpp>
+#include <extras/rsi/subsystems.hpp>
 #include <fstream>
 #include <iostream>
 #include <sstream>
@@ -10,9 +10,9 @@ using namespace std;
 namespace extras {
 
   RequestedService Services::request(const PortNumber &serverSocket) {
-    RequestedService serviceName = read_line(serverSocket);
+    RequestedService serviceName = extras::rsi::read_line(serverSocket);
     _lastPortRequested = _portAuthority.request();
-    send_line(to_string(_lastPortRequested), serverSocket);
+    extras::rsi::send_line(to_string(_lastPortRequested), serverSocket);
     return serviceName;
   }